moved lightning extension to nano/pytorch
